+++ R/D_regularized_fold_out.R
@@ -97,7 +97,7 @@ D_regularized_fold_out <-
 
     train.data <- dplyr::sample_n(data.grouped,
       size = size,
-      replace = F
+      replace = FALSE
     )
 
     test.data <- data[!(data$row.nmbr %in% train.data$row.nmbr), ]