+++ R/D_regularized_out.R
@@ -14,7 +14,7 @@
 #' @param auc Logical. Include area under the receiver operating characteristics? Default FALSE.
 #' @param pred.prob Logical. Include table of predicted probabilities? Default FALSE.
 #' @param prob.cutoffs Vector. Cutoffs for table of predicted probabilities. Default seq(0,1,0.20).
-#'
+#' @param append.data Logical. If TRUE, the testing data split is appended to the predicted variables.
 #' @return
 #' \item{D}{Multivariate descriptive statistics and differences.}
 #' \item{pred.dat}{A data.frame with predicted values.}
@@ -48,7 +48,8 @@ D_regularized_out <-
            pcc = FALSE,
            auc = FALSE,
            pred.prob = FALSE,
-           prob.cutoffs = seq(from = 0, to = 1, by = 0.20)) {
+           prob.cutoffs = seq(from = 0, to = 1, by = 0.20),
+           append.data = FALSE) {
     data$group.var.num <-
       ifelse(data[, group.var] == group.values[1], 1,
         ifelse(data[, group.var] == group.values[2], 0,
@@ -93,6 +94,10 @@ D_regularized_out <-
         )
       )
     )
+
+    if (append.data) {
+      preds <- cbind(preds, test.data)
+    }
 
     D <- multid::d_pooled_sd(
       data = preds,