+++ NEWS.md
@@ -1,4 +1,4 @@
-# multid 0.5.0.9000
+# multid 0.6.0
 
 * Bug fixed in D_regularized_out
 * Addend.data argument added to all D_regularized -functions. In _fold -functions, test-partition of the data is appended, else the entire data frame is added.